ProASP.NETMVC5/SportsStore.WebUI/Controllers/AccountController.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Web;
using System.Web.Mvc;

using SportsStore.WebUI.Infrastructure.Abstract;
using SportsStore.WebUI.Models;

namespace SportsStore.WebUI.Controllers
{
    public class AccountController : Controller
    {
        private IAuthenticationProvider m_authenticationProvider;

        public AccountController(IAuthenticationProvider authenticationProvider)
        {
            m_authenticationProvider = authenticationProvider;
        }

        public ActionResult Login()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Login(LoginViewModel viewModel, String returnUrl)
        {
            if (ModelState.IsValid)
            {
                if (m_authenticationProvider.Authenticate(viewModel.UserName, viewModel.Password))
                {
                    return Redirect(returnUrl ?? Url.Action("Index", "Admin"));
                }
                else
                {
                    ModelState.AddModelError("", "Incorrect username or password");
                    return View();
                }
            }
            else
            {
                return View();
            }
        }
    }
}